# docs/openapi/order-service.json
{
  "components": {
    "schemas": {
      "ActivityOut": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "bid_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid Id"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The seller (null for the platform) of a SALE; the owner of a listing",
            "title": "From"
          },
          "kind": {
            "enum": [
              "SALE",
              "LISTED",
              "UNLISTED",
              "BID",
              "BID_WITHDRAWN"
            ],
            "title": "Kind",
            "type": "string"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The buyer of a SALE; the bidder of a bid",
            "title": "To"
          }
        },
        "required": [
          "kind",
          "sku",
          "at",
          "amount",
          "currency",
          "from",
          "to",
          "order_id",
          "bid_id"
        ],
        "title": "ActivityOut",
        "type": "object"
      },
      "ActivityPageOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActivityOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "size",
          "total"
        ],
        "title": "ActivityPageOut",
        "type": "object"
      },
      "BidAccept": {
        "additionalProperties": false,
        "properties": {
          "customer_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
            "title": "Customer Id",
            "type": "string"
          }
        },
        "required": [
          "customer_id"
        ],
        "title": "BidAccept",
        "type": "object"
      },
      "BidCreate": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "examples": [
              "33.50"
            ],
            "pattern": "^[0-9]{1,8}(\\.[0-9]{1,2})?$",
            "title": "Amount",
            "type": "string"
          },
          "customer_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
            "title": "Customer Id",
            "type": "string"
          },
          "sku": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
            "title": "Sku",
            "type": "string"
          }
        },
        "required": [
          "customer_id",
          "sku",
          "amount"
        ],
        "title": "BidCreate",
        "type": "object"
      },
      "BidOut": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "bid_id": {
            "title": "Bid Id",
            "type": "string"
          },
          "bidder": {
            "title": "Bidder",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "listing_id": {
            "title": "Listing Id",
            "type": "string"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "status": {
            "enum": [
              "OPEN",
              "WITHDRAWN",
              "ACCEPTED",
              "FILLED",
              "FAILED",
              "CLOSED"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "bid_id",
          "listing_id",
          "sku",
          "bidder",
          "amount",
          "currency",
          "status",
          "order_id",
          "created_at",
          "updated_at"
        ],
        "title": "BidOut",
        "type": "object"
      },
      "BidPageOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BidOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "size",
          "total"
        ],
        "title": "BidPageOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ItemOut": {
        "properties": {
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "seller": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Who it was bought from: null for the platform, else the previous owner",
            "title": "Seller"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "unit_price": {
            "title": "Unit Price",
            "type": "string"
          }
        },
        "required": [
          "sku",
          "quantity",
          "unit_price",
          "seller"
        ],
        "title": "ItemOut",
        "type": "object"
      },
...
      "ListingCreate": {
        "additionalProperties": false,
        "properties": {
          "customer_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
            "title": "Customer Id",
            "type": "string"
          },
          "sku": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
            "title": "Sku",
            "type": "string"
          }
        },
        "required": [
          "customer_id",
          "sku"
        ],
        "title": "ListingCreate",
        "type": "object"
      },
      "ListingOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "listing_id": {
            "title": "Listing Id",
            "type": "string"
          },
          "seller": {
            "title": "Seller",
            "type": "string"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "status": {
            "enum": [
              "OPEN",
              "SALE_PENDING",
              "SOLD",
              "CANCELLED"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "listing_id",
          "sku",
          "seller",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ListingOut",
        "type": "object"
      },
      "ListingPageOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ListingOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "size",
          "total"
        ],
        "title": "ListingPageOut",
        "type": "object"
      },
...
      "OrderOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "order_id": {
            "title": "Order Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "total_amount": {
            "title": "Total Amount",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "order_id",
          "customer_id",
          "status",
          "status_reason",
          "total_amount",
          "currency",
          "items",
          "created_at",
          "updated_at"
        ],
        "title": "OrderOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "order-service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/activity": {
      "get": {
        "description": "Sales, listings and bids, newest first; for one CloudPunk with ``sku``.",
        "operationId": "activity_api_v1_activity_get",
        "parameters": [
          {
            "in": "query",
            "name": "sku",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sku"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activity"
      }
    },
    "/api/v1/bids": {
      "get": {
        "description": "Bids on a CloudPunk (``sku``) or by a customer (``customer_id``); at least one is needed.",
        "operationId": "list_bids_api_v1_bids_get",
        "parameters": [
          {
            "in": "query",
            "name": "sku",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sku"
            }
          },
          {
            "in": "query",
            "name": "customer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "OPEN",
                    "WITHDRAWN",
                    "ACCEPTED",
                    "FILLED",
                    "FAILED",
                    "CLOSED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BidPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Bids"
      },
      "post": {
        "description": "A bid on a CloudPunk that is up for bid. The same key and body again returns it (200).",
        "operationId": "place_bid_api_v1_bids_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9._:-]{1,64}$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BidCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BidOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Place Bid"
      }
    },
    "/api/v1/bids/{bid_id}": {
      "delete": {
        "operationId": "withdraw_bid_api_v1_bids__bid_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "bid_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Bid Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "customer_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
              "title": "Customer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BidOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Withdraw Bid"
      }
    },
    "/api/v1/bids/{bid_id}/accept": {
      "post": {
        "description": "The owner accepts a bid: 202 with the bidder's order, PENDING until inventory moves the\nCloudPunk. Accepting it again returns the same order (200).",
        "operationId": "accept_bid_api_v1_bids__bid_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "bid_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Bid Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BidAccept"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Bid"
      }
    },
    "/api/v1/listings": {
      "get": {
        "description": "``status=ACTIVE`` (the default) is the purple set: open, or with a bid being confirmed.",
        "operationId": "list_listings_api_v1_listings_get",
        "parameters": [
          {
            "in": "query",
            "name": "sku",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sku"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "ACTIVE",
              "enum": [
                "ACTIVE",
                "OPEN",
                "SALE_PENDING",
                "SOLD",
                "CANCELLED",
                "ALL"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Listings"
      },
      "post": {
        "description": "The owner puts a CloudPunk up for bid (it turns purple). Already up: 200 with that one.",
        "operationId": "put_up_api_v1_listings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Up"
      }
    },
    "/api/v1/listings/{sku}": {
      "delete": {
        "description": "The owner takes it off the market (it turns blue); its open bids close.",
        "operationId": "take_off_api_v1_listings__sku__delete",
        "parameters": [
          {
            "in": "path",
            "name": "sku",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$",
              "title": "Sku",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "customer_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
              "title": "Customer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Take Off"
      }
    },
...
    "/api/v1/orders/{order_id}": {
      "get": {
        "operationId": "get_order_api_v1_orders__order_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$",
              "title": "Order Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Order"
      }
    },
...
  }
}